+++ unikmer/cmd/common.go
@@ -26,11 +26,12 @@ import (
 	"io"
 	"os"
 	"runtime"
-	"sort"
 
 	"github.com/pkg/errors"
 	"github.com/shenwei356/unikmer"
 	"github.com/spf13/cobra"
+	"github.com/twotwotwo/sorts"
+	"github.com/twotwotwo/sorts/sortutil"
 )
 
 var commonCmd = &cobra.Command{
@@ -57,6 +58,7 @@ Tips:
 	Run: func(cmd *cobra.Command, args []string) {
 		opt := getOptions(cmd)
 		runtime.GOMAXPROCS(opt.NumCPUs)
+		sorts.MaxProcs = opt.NumCPUs
 
 		var err error
 
@@ -339,7 +341,8 @@ Tips:
 			log.Infof("no shared k-mers found")
 		}
 
-		sort.Sort(unikmer.CodeSlice(codes))
+		// sort.Sort(unikmer.CodeSlice(codes))
+		sortutil.Uint64s(codes)
 
 		if hasTaxid || hasMixTaxid {
 			for _, code := range codes {