+++ unikmer/cmd/split.go
@@ -27,11 +27,12 @@ import (
 	"os"
 	"path/filepath"
 	"runtime"
-	"sort"
 	"sync"
 
 	"github.com/pkg/errors"
 	"github.com/shenwei356/util/pathutil"
+	"github.com/twotwotwo/sorts"
+	"github.com/twotwotwo/sorts/sortutil"
 
 	"github.com/shenwei356/unikmer"
 	"github.com/spf13/cobra"
@@ -57,6 +58,7 @@ Tips:
 	Run: func(cmd *cobra.Command, args []string) {
 		opt := getOptions(cmd)
 		runtime.GOMAXPROCS(opt.NumCPUs)
+		sorts.MaxProcs = opt.NumCPUs
 
 		outDir := getFlagString(cmd, "out-dir")
 		force := getFlagBool(cmd, "force")
@@ -300,12 +302,14 @@ Tips:
 								if opt.Verbose {
 									log.Infof("[chunk %d] sorting %d k-mers", iTmpFile, len(mt))
 								}
-								sort.Sort(unikmer.CodeTaxidSlice(mt))
+								// sort.Sort(unikmer.CodeTaxidSlice(mt))
+								sorts.Quicksort(unikmer.CodeTaxidSlice(mt))
 							} else {
 								if opt.Verbose {
 									log.Infof("[chunk %d] sorting %d k-mers", iTmpFile, len(m))
 								}
-								sort.Sort(unikmer.CodeSlice(m))
+								// sort.Sort(unikmer.CodeSlice(m))
+								sortutil.Uint64s(m)
 							}
 
 							var _n int64
@@ -376,7 +380,8 @@ Tips:
 				if opt.Verbose {
 					log.Infof("[chunk %d] sorting %d k-mers", iTmpFile, len(m))
 				}
-				sort.Sort(unikmer.CodeSlice(m))
+				// sort.Sort(unikmer.CodeSlice(m))
+				sortutil.Uint64s(m)
 				if opt.Verbose {
 					log.Infof("[chunk %d] done sorting", iTmpFile)
 					log.Infof("[chunk %d] writing to file: %s", iTmpFile, outFile)